--- src/main/java/was/example/domain/PartnerStore.java
@@ -1,5 +1,0 @@
-package was.example.domain;
-
-public interface PartnerStore {
-    Partner store(Partner initPartner);
-}
