--- src/main/java/was/example/facade/PartnerFacade.java
@@ -1,23 +1,0 @@
-package was.example.facade;
-
-import was.notification.application.NotificationService;
-import was.example.domain.PartnerCommand;
-import was.example.domain.PartnerInfo;
-import was.example.domain.PartnerService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class PartnerFacade {
-    private final PartnerService partnerService;
-    private final NotificationService notificationService;
-
-    public PartnerInfo registerPartner(PartnerCommand command) {
-        var partnerInfo = partnerService.registerPartner(command);
-        notificationService.sendEmail(partnerInfo.getEmail(), "title", "description");
-        return partnerInfo;
-    }
-}
