+++ src/main/java/was/examplePakege/example/domain/Member/Member.java
@@ -1,0 +1,4 @@
+package was.examplePakege.example.domain.Member;
+
+public class Member {
+}
