+++ src/main/java/was/examplePakege/example/domain/partner/PartnerService.java
@@ -1,4 +1,4 @@
-package was.example.domain;
+package was.examplePakege.example.domain.partner;
 
 public interface PartnerService {
     PartnerInfo registerPartner(PartnerCommand command);
