+++ src/main/java/was/examplePakege/example/infrastructure/Store/PartnerStoreImpl.java
@@ -1,9 +1,9 @@
-package was.example.infrastructure.Store;
+package was.examplePakege.example.infrastructure.Store;
 
 import was.common.exception.InvalidParamException;
-import was.example.domain.Partner;
-import was.example.domain.PartnerStore;
-import was.example.infrastructure.PartnerRepository;
+import was.examplePakege.example.domain.partner.Partner;
+import was.examplePakege.example.domain.partner.PartnerStore;
+import was.examplePakege.example.infrastructure.PartnerRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
