+++ src/main/java/was/examplePakege/examplenotification/application/NotificationService.java
@@ -1,4 +1,4 @@
-package was.notification.application;
+package was.examplePakege.examplenotification.application;
 
 public interface NotificationService {
     void sendEmail(String email, String title, String description);
